+++ PENDING.md
@@ -43,6 +43,8 @@ BUG FIXES
 * Gaia REST API
 
 * Gaia CLI
+  * [\#3586](https://github.com/cosmos/cosmos-sdk/pull/3586) Incomplete ledger derivation paths in keybase 
+
 
 * Gaia
