+++ crypto/keys/codec.go
@@ -11,7 +11,7 @@ var cdc = amino.NewCodec()
 func init() {
 	cryptoAmino.RegisterAmino(cdc)
 	cdc.RegisterInterface((*Info)(nil), nil)
-	cdc.RegisterConcrete(hd.BIP44Params{},"crypto/keys/hd/BIP44Params", nil)
+	cdc.RegisterConcrete(hd.BIP44Params{}, "crypto/keys/hd/BIP44Params", nil)
 	cdc.RegisterConcrete(localInfo{}, "crypto/keys/localInfo", nil)
 	cdc.RegisterConcrete(ledgerInfo{}, "crypto/keys/ledgerInfo", nil)
 	cdc.RegisterConcrete(offlineInfo{}, "crypto/keys/offlineInfo", nil)
